requirement.txt
asgiref==3.8.1
bcrypt==4.1.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.7
Django==4.1.13
django-cron==0.6.0
django-filter
djangorestframework==3.15.1
djongo==1.3.6
dnspython==2.6.1
ecdsa==0.11
idna==3.7
Markdown==3.6
paramiko==3.4.0
pbr==6.0.0
pycparser==2.22
pymongo==3.12.3
PyNaCl==1.5.0
python-gerrit-api==3.0.8
pytz==2024.1
requests==2.32.3
sqlparse==0.2.4
typing_extensions==4.12.0
urllib3==2.2.1
retry==0.9.2
xmltodict==0.13.0
nltk==3.8.1
beautifulsoup4==4.12.3
unparallel==0.4.0